requirements.txt
pypokerengine==1.0.1
numpy==1.24.0
scipy==1.10.0
scikit-learn==1.3.0
joblib==1.3.0
pyyaml==6.0
tqdm==4.65.0